binexport: init refactor for multi-arch instruction feature parsing
