+++ capa/features/extractors/binexport2/insn.py
@@ -6,13 +6,15 @@
 #  is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 import logging
-from typing import List, Tuple, Iterator, Optional
+from typing import List, Tuple, Iterator
 
 import capa.features.extractors.helpers
 import capa.features.extractors.strings
 import capa.features.extractors.binexport2.helpers
-from capa.features.insn import API, MAX_STRUCTURE_SIZE, Number, Offset, Mnemonic, OperandNumber, OperandOffset
-from capa.features.common import ARCH_I386, ARCH_AMD64, ARCH_AARCH64, Bytes, String, Feature, Characteristic
+import capa.features.extractors.binexport2.arch.arm.insn
+import capa.features.extractors.binexport2.arch.intel.insn
+from capa.features.insn import API, Mnemonic
+from capa.features.common import Bytes, String, Feature, Characteristic
 from capa.features.address import Address, AbsoluteVirtualAddress
 from capa.features.extractors.binexport2 import (
     AddressSpace,
@@ -20,39 +22,14 @@ from capa.features.extractors.binexport2 import (
     BinExport2Index,
     FunctionContext,
     ReadMemoryError,
-    BasicBlockContext,
     BinExport2Analysis,
     InstructionContext,
 )
 from capa.features.extractors.base_extractor import BBHandle, InsnHandle, FunctionHandle
-from capa.features.extractors.binexport2.helpers import (
-    OperandPhraseInfo,
-    get_operand_phrase_info,
-    get_operand_register_expression,
-    get_operand_immediate_expression,
-)
+from capa.features.extractors.binexport2.helpers import HAS_ARCH_ARM, HAS_ARCH_INTEL
 from capa.features.extractors.binexport2.binexport2_pb2 import BinExport2
 
 logger = logging.getLogger(__name__)
-
-# security cookie checks may perform non-zeroing XORs, these are expected within a certain
-# byte range within the first and returning basic blocks, this helps to reduce FP features
-SECURITY_COOKIE_BYTES_DELTA: int = 0x40
-
-
-HAS_ARCH32 = {ARCH_I386}
-HAS_ARCH64 = {ARCH_AARCH64, ARCH_AMD64}
-
-HAS_ARCH_INTEL = {ARCH_I386, ARCH_AMD64}
-HAS_ARCH_ARM = {ARCH_AARCH64}
-
-
-def mask_immediate(fhi: FunctionContext, immediate: int) -> int:
-    if fhi.arch & HAS_ARCH64:
-        immediate &= 0xFFFFFFFFFFFFFFFF
-    elif fhi.arch & HAS_ARCH32:
-        immediate &= 0xFFFFFFFF
-    return immediate
 
 
 def extract_insn_api_features(fh: FunctionHandle, _bbh: BBHandle, ih: InsnHandle) -> Iterator[Tuple[Feature, Address]]:
@@ -103,65 +80,15 @@ def extract_insn_api_features(fh: FunctionHandle, _bbh: BBHandle, ih: InsnHandle
     """
 
 
-def is_address_mapped(be2: BinExport2, address: int) -> bool:
-    """return True if the given address is mapped"""
-    sections_with_perms: Iterator[BinExport2.Section] = filter(lambda s: s.flag_r or s.flag_w or s.flag_x, be2.section)
-    return any(section.address <= address < section.address + section.size for section in sections_with_perms)
-
-
 def extract_insn_number_features(
-    fh: FunctionHandle, _bbh: BBHandle, ih: InsnHandle
+    fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle
 ) -> Iterator[Tuple[Feature, Address]]:
     fhi: FunctionContext = fh.inner
-    ii: InstructionContext = ih.inner
-
-    be2: BinExport2 = fhi.ctx.be2
-
-    instruction_index: int = ii.instruction_index
-    instruction: BinExport2.Instruction = be2.instruction[instruction_index]
-
-    if len(instruction.operand_index) == 0:
-        # skip things like:
-        #   .text:0040116e leave
-        return
-
-    mnemonic: str = be2.mnemonic[instruction.mnemonic_index].name.lower()
 
     if fhi.arch & HAS_ARCH_INTEL:
-        # short-circut checks for intel architecture
-        if mnemonic.startswith("ret"):
-            # skip things like:
-            #   .text:0042250E retn 8
-            return
-
-        if mnemonic.startswith(("add", "sub")):
-            register_expression: Optional[BinExport2.Expression] = get_operand_register_expression(
-                be2, be2.operand[instruction.operand_index[0]]
-            )
-            if register_expression and register_expression.symbol.lower().endswith(("sp", "bp")):
-                # skip things like:
-                # 0x415bbc  ADD         ESP, 0xC
-                return
-
-    for i, operand_index in enumerate(instruction.operand_index):
-        operand: BinExport2.Operand = be2.operand[operand_index]
-
-        immediate_expression: Optional[BinExport2.Expression] = get_operand_immediate_expression(be2, operand)
-        if not immediate_expression:
-            continue
-
-        value: int = mask_immediate(fhi, immediate_expression.immediate)
-        if is_address_mapped(be2, value):
-            continue
-
-        yield Number(value), ih.address
-        yield OperandNumber(i, value), ih.address
-
-        if fhi.arch & HAS_ARCH_INTEL:
-            if mnemonic.startswith("add"):
-                if 0 < value < MAX_STRUCTURE_SIZE:
-                    yield Offset(value), ih.address
-                    yield OperandOffset(i, value), ih.address
+        yield from capa.features.extractors.binexport2.arch.intel.insn.extract_insn_number_features(fh, bbh, ih)
+    elif fhi.arch & HAS_ARCH_ARM:
+        yield from capa.features.extractors.binexport2.arch.arm.insn.extract_insn_number_features(fh, bbh, ih)
 
 
 def extract_insn_bytes_features(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle) -> Iterator[Tuple[Feature, Address]]:
@@ -246,133 +173,26 @@ def extract_insn_offset_features(
     fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle
 ) -> Iterator[Tuple[Feature, Address]]:
     fhi: FunctionContext = fh.inner
-    ii: InstructionContext = ih.inner
 
-    be2: BinExport2 = fhi.ctx.be2
-    instruction: BinExport2.Instruction = be2.instruction[ii.instruction_index]
-
-    if len(instruction.operand_index) == 0:
-        # skip things like:
-        #   .text:0040116e leave
-        return
-
-    mnemonic: str = be2.mnemonic[instruction.mnemonic_index].name.lower()
-
-    for i, operand_index in enumerate(instruction.operand_index):
-        operand: BinExport2.Operand = be2.operand[operand_index]
-
-        is_dereference = False
-        for expression_index in operand.expression_index:
-            if be2.expression[expression_index].type == BinExport2.Expression.DEREFERENCE:
-                is_dereference = True
-                break
-
-        if not is_dereference:
-            continue
-
-        if fhi.arch & HAS_ARCH_INTEL:
-            phrase_info: Optional[OperandPhraseInfo] = get_operand_phrase_info(be2, operand)
-            if not phrase_info:
-                continue
-
-            if phrase_info.displacement:
-                if phrase_info.base and phrase_info.base.symbol.lower().endswith(("bp", "sp")):
-                    # skips things like:
-                    # 00401068 MOV dword ptr [EBP + local_8],EAX
-                    continue
-
-                value: int = mask_immediate(fhi, phrase_info.displacement.immediate)
-                if not is_address_mapped(be2, value):
-                    value = capa.features.extractors.helpers.twos_complement(value, 32)
-
-                    yield Offset(value), ih.address
-                    yield OperandOffset(i, value), ih.address
-
-                    if mnemonic == "lea" and i == 1:
-                        if phrase_info.base and not any((phrase_info.scale, phrase_info.index)):
-                            yield Number(value), ih.address
-                            yield OperandNumber(i, value), ih.address
-
-            elif phrase_info.base and not any((phrase_info.index, phrase_info.scale)):
-                # like:
-                # 00401062 MOVZX EAX,word ptr [EDI]
-                yield Offset(0), ih.address
-                yield OperandOffset(i, 0), ih.address
-
-
-def is_security_cookie(
-    fhi: FunctionContext,
-    bbi: BasicBlockContext,
-    instruction: BinExport2.Instruction,
-) -> bool:
-    """
-    check if an instruction is related to security cookie checks.
-    """
-    be2: BinExport2 = fhi.ctx.be2
-
-    # security cookie check should use SP or BP
-    op1: BinExport2.Operand = be2.operand[instruction.operand_index[1]]
-    op1_exprs: List[BinExport2.Expression] = [be2.expression[expr_i] for expr_i in op1.expression_index]
-    if all(expr.symbol.lower() not in ("bp", "esp", "ebp", "rbp", "rsp") for expr in op1_exprs):
-        return False
-
-    # check_nzxor_security_cookie_delta
-    # if insn falls at the start of first entry block of the parent function.
-    flow_graph: BinExport2.FlowGraph = be2.flow_graph[fhi.flow_graph_index]
-    basic_block_index: int = bbi.basic_block_index
-    bb: BinExport2.BasicBlock = be2.basic_block[basic_block_index]
-    if flow_graph.entry_basic_block_index == basic_block_index:
-        first_addr: int = min((be2.instruction[ir.begin_index].address for ir in bb.instruction_index))
-        if instruction.address < first_addr + SECURITY_COOKIE_BYTES_DELTA:
-            return True
-    # or insn falls at the end before return in a terminal basic block.
-    if basic_block_index not in (e.source_basic_block_index for e in flow_graph.edge):
-        last_addr: int = max((be2.instruction[ir.end_index - 1].address for ir in bb.instruction_index))
-        if instruction.address > last_addr - SECURITY_COOKIE_BYTES_DELTA:
-            return True
-    return False
+    if fhi.arch & HAS_ARCH_INTEL:
+        yield from capa.features.extractors.binexport2.arch.intel.insn.extract_insn_offset_features(fh, bbh, ih)
+    elif fhi.arch & HAS_ARCH_ARM:
+        yield from capa.features.extractors.binexport2.arch.arm.insn.extract_insn_offset_features(fh, bbh, ih)
 
 
 def extract_insn_nzxor_characteristic_features(
     fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle
 ) -> Iterator[Tuple[Feature, Address]]:
-    """
-    parse non-zeroing XOR instruction from the given instruction.
-    ignore expected non-zeroing XORs, e.g. security cookies.
-    """
     fhi: FunctionContext = fh.inner
-    ii: InstructionContext = ih.inner
 
-    be2: BinExport2 = fhi.ctx.be2
-
-    instruction: BinExport2.Instruction = be2.instruction[ii.instruction_index]
-    mnemonic: BinExport2.Mnemonic = be2.mnemonic[instruction.mnemonic_index]
-    mnemonic_name: str = mnemonic.name.lower()
-    if mnemonic_name not in (
-        "xor",
-        "xorpd",
-        "xorps",
-        "pxor",  # x86 / amd64
-        "eor",  # arm / aarch64
-    ):
-        return
-
-    operands: List[BinExport2.Operand] = [be2.operand[operand_index] for operand_index in instruction.operand_index]
-
-    # check whether operands are same for x86 / amd64
-    if mnemonic_name in ("xor", "xorpd", "xorps", "pxor"):
-        if operands[0] == operands[1]:
-            return
-        if is_security_cookie(fhi, bbh.inner, instruction):
-            return
-
-    # check whether 2nd/3rd operands are same for arm / aarch64
-    if mnemonic_name == "eor":
-        assert len(operands) == 3
-        if operands[1] == operands[2]:
-            return
-
-    yield Characteristic("nzxor"), ih.address
+    if fhi.arch & HAS_ARCH_INTEL:
+        yield from capa.features.extractors.binexport2.arch.intel.insn.extract_insn_nzxor_characteristic_features(
+            fh, bbh, ih
+        )
+    elif fhi.arch & HAS_ARCH_ARM:
+        yield from capa.features.extractors.binexport2.arch.arm.insn.extract_insn_nzxor_characteristic_features(
+            fh, bbh, ih
+        )
 
 
 def extract_insn_mnemonic_features(
@@ -412,9 +232,16 @@ def extract_function_calls_from(fh: FunctionHandle, bbh: BBHandle, ih: InsnHandl
 def extract_function_indirect_call_characteristic_features(
     fh: FunctionHandle, bbh: BBHandle, ih: InsnHandle
 ) -> Iterator[Tuple[Feature, Address]]:
-    # TODO(wb): complete
-    # https://github.com/mandiant/capa/issues/1755
-    yield from ()
+    fhi: FunctionContext = fh.inner
+
+    if fhi.arch & HAS_ARCH_INTEL:
+        yield from capa.features.extractors.binexport2.arch.intel.insn.extract_function_indirect_call_characteristic_features(
+            fh, bbh, ih
+        )
+    elif fhi.arch & HAS_ARCH_ARM:
+        yield from capa.features.extractors.binexport2.arch.arm.insn.extract_function_indirect_call_characteristic_features(
+            fh, bbh, ih
+        )
 
 
 def extract_features(f: FunctionHandle, bbh: BBHandle, insn: InsnHandle) -> Iterator[Tuple[Feature, Address]]:
